+++ src/Symfony/Component/DependencyInjection/Tests/Compiler/MergeExtensionConfigurationPassTest.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\DependencyInjection\Tests\Compiler;
 
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\Attributes\IgnoreDeprecations;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
@@ -18,6 +20,8 @@ use Symfony\Component\Config\Resource\FileResource;
 use Symfony\Component\DependencyInjection\Compiler\MergeExtensionConfigurationContainerBuilder;
 use Symfony\Component\DependencyInjection\Compiler\MergeExtensionConfigurationPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
+use Symfony\Component\DependencyInjection\Exception\LogicException;
 use Symfony\Component\DependencyInjection\Exception\ParameterNotFoundException;
 use Symfony\Component\DependencyInjection\Exception\RuntimeException;
 use Symfony\Component\DependencyInjection\Extension\Extension;
@@ -72,7 +76,7 @@ class MergeExtensionConfigurationPassTest extends TestCase
     public function testExtensionConfigurationIsTrackedByDefault()
     {
         $extension = $this->getMockBuilder(FooExtension::class)->onlyMethods(['getConfiguration'])->getMock();
-        $extension->expects($this->exactly(2))
+        $extension->expects($this->exactly(3))
             ->method('getConfiguration')
             ->willReturn(new FooConfiguration());
 
@@ -158,6 +162,89 @@ class MergeExtensionConfigurationPassTest extends TestCase
         (new MergeExtensionConfigurationPass())->process($container);
 
         $this->addToAssertionCount(1);
+    }
+
+    public function testExtensionAliasesAreForwardedToTheAliasedExtension()
+    {
+        $container = new ContainerBuilder();
+        $container->registerExtension(new AliasingExtension());
+        $container->registerExtension(new TargetExtension());
+        $container->loadFromExtension('aliasing', ['own' => 'first', 'target' => ['value' => 'forwarded first']]);
+        $container->loadFromExtension('aliasing', ['target' => ['value' => 'forwarded second']]);
+        $container->loadFromExtension('target', ['value' => 'explicit']);
+
+        (new MergeExtensionConfigurationPass())->process($container);
+
+        $this->assertSame([['own' => 'first'], []], $container->getParameter('aliasing.configs'));
+        $this->assertSame([['value' => 'forwarded first'], ['value' => 'forwarded second'], ['value' => 'explicit']], $container->getParameter('target.configs'));
+        $this->assertSame([['own' => 'first'], []], $container->getExtensionConfig('aliasing'));
+    }
+
+    public function testExtensionAliasesNormalizeTheirValueBeforeForwarding()
+    {
+        $container = new ContainerBuilder();
+        $container->registerExtension(new AliasingExtension());
+        $container->registerExtension(new TargetExtension());
+        $container->loadFromExtension('aliasing', ['target' => null]);
+        $container->loadFromExtension('aliasing', ['target' => false]);
+
+        (new MergeExtensionConfigurationPass())->process($container);
+
+        $this->assertSame([[], ['enabled' => false]], $container->getParameter('target.configs'));
+    }
+
+    public function testExtensionAliasesUseTheXmlRemappingOfTheirParent()
+    {
+        $container = new ContainerBuilder();
+        $container->registerExtension(new AliasingExtension());
+        $container->registerExtension(new TargetExtension());
+        $container->loadFromExtension('aliasing', ['item' => ['value' => 'single']]);
+        $container->loadFromExtension('aliasing', ['item' => [['value' => 'first'], ['value' => 'second']]]);
+
+        (new MergeExtensionConfigurationPass())->process($container);
+
+        $this->assertSame([[['value' => 'single']], [['value' => 'first'], ['value' => 'second']]], $container->getParameter('target.configs'));
+    }
+
+    public function testExtensionAliasesRejectValuesThatAreNotArrays()
+    {
+        $container = new ContainerBuilder();
+        $container->registerExtension(new AliasingExtension());
+        $container->registerExtension(new TargetExtension());
+        $container->loadFromExtension('aliasing', ['target' => 'not an array']);
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The "aliasing.target" configuration must be an array or null, "string" given.');
+
+        (new MergeExtensionConfigurationPass())->process($container);
+    }
+
+    public function testExtensionAliasesRequireTheAliasedExtension()
+    {
+        $container = new ContainerBuilder();
+        $container->registerExtension(new AliasingExtension());
+        $container->loadFromExtension('aliasing', ['target' => []]);
+
+        $this->expectException(LogicException::class);
+        $this->expectExceptionMessage('The "aliasing.target" configuration is handled by the "target" extension, which is not registered.');
+
+        (new MergeExtensionConfigurationPass())->process($container);
+    }
+
+    #[Group('legacy')]
+    #[IgnoreDeprecations]
+    public function testDeprecatedExtensionAliasesTriggerTheirDeprecation()
+    {
+        $container = new ContainerBuilder();
+        $container->registerExtension(new AliasingExtension());
+        $container->registerExtension(new TargetExtension());
+        $container->loadFromExtension('aliasing', ['legacy' => ['value' => 'legacy']]);
+
+        $this->expectUserDeprecationMessage('Since symfony/test 1.0: The "aliasing.legacy" configuration is deprecated, use "target" instead.');
+
+        (new MergeExtensionConfigurationPass())->process($container);
+
+        $this->assertSame([['value' => 'legacy']], $container->getParameter('target.configs'));
     }
 }
 
@@ -259,5 +346,54 @@ final class TestCccExtension extends Extension
     {
         $configuration = $this->getConfiguration($configs, $container);
         $this->processConfiguration($configuration, $configs);
+    }
+}
+
+final class AliasingConfiguration implements ConfigurationInterface
+{
+    public function getConfigTreeBuilder(): TreeBuilder
+    {
+        $treeBuilder = new TreeBuilder('aliasing');
+        $treeBuilder->getRootNode()
+            ->fixXmlConfig('item')
+            ->children()
+                ->scalarNode('own')->end()
+                ->variableNode('target')->extension('target')->treatFalseLike(['enabled' => false])->end()
+                ->variableNode('items')->extension('target')->end()
+                ->variableNode('legacy')->extension('target')->setDeprecated('symfony/test', '1.0', 'The "%path%" configuration is deprecated, use "target" instead.')->end()
+            ->end();
+
+        return $treeBuilder;
+    }
+}
+
+final class AliasingExtension extends Extension
+{
+    public function getAlias(): string
+    {
+        return 'aliasing';
+    }
+
+    public function getConfiguration(array $config, ContainerBuilder $container): ?ConfigurationInterface
+    {
+        return new AliasingConfiguration();
+    }
+
+    public function load(array $configs, ContainerBuilder $container): void
+    {
+        $container->setParameter('aliasing.configs', $configs);
+    }
+}
+
+final class TargetExtension extends Extension
+{
+    public function getAlias(): string
+    {
+        return 'target';
+    }
+
+    public function load(array $configs, ContainerBuilder $container): void
+    {
+        $container->setParameter('target.configs', $configs);
     }
 }